Add new batch item functionality to back-end - plus refactor of int to long for ids from controller to service to repo;

+++ Data/Queries/BatchItemRepository.cs
@@ -13,7 +13,7 @@ namespace Data.Queries
         {
         }
 
-        public async Task<IEnumerable<BatchItem>> Get(int batchId)
+        public async Task<IEnumerable<BatchItem>> Get(long batchId)
         {
             return await _context.BatchItems
                 .Include(x => x.Batch)
